+++ 02_input_output/input_output.py
@@ -1,0 +1,11 @@
+'''
+'''
+
+#input and output : input() and print() functions
+name = input("Enter your name: ")  # input() function to take input from user
+collegeName = input("Enter your college name: ")
+print("Hello, " + name + "!")  # print() function to display output
+print("Welcome to " + collegeName + ".")
+
+
+